refactor: entitiesとfeaturesのui/index.tsファイルを削除
- src/entities/post/ui/index.tsを削除
- src/features/follow/ui/index.tsを削除
- 上位のindex.tsファイルで直接コンポーネントをエクスポート
- FSDアーキテクチャの推奨構造に準拠
