+++ src/entities/post/index.ts
@@ -1,1 +1,2 @@
-export { Post, PostActionMenu } from "./ui";
+export { Post } from "./ui/Post";
+export { PostActionMenu } from "./ui/PostActionMenu";
