--- src/entities/post/ui/index.ts
@@ -1,1 +1,0 @@
-export { Post } from "./Post";
