--- src/features/follow/ui/index.ts
@@ -1,1 +1,0 @@
-export { FollowButton } from "./FollowButton";
